Use hermetic Python. (#138)
The currently specified versions of scipy and numpy no longer work with
Python 3.10, but their latest versions are not available for Python 3.6,
which is the version available on the Ubuntu 18.04 CI machines.

Using a hermetic interpreter ensures builds are reproducible irrespective
of the interpreter version available on the host machine.

Also note that it's not necessary to run pip on the CI machines. The
Python rules are already able to fetch the dependencies.